pets_classifer_system/requirements.txt
Django==4.0.4
djangorestframework==3.13.1
tensorflow==2.12.0
tensorflow_intel==2.12.0
